refactored data model
