+++ app/src/main/java/com/kickstarter/libs/utils/SharedFunctions.kt
@@ -1,12 +1,8 @@
 package com.kickstarter.libs.utils
 
-import com.google.android.gms.common.util.Base64Utils
-import com.kickstarter.models.Relay
 import com.kickstarter.models.Reward
 import com.kickstarter.ui.data.CheckoutData
 import type.CreditCardPaymentType
-import java.nio.charset.Charset
-import kotlin.math.absoluteValue
 
 /**
  * Total count of selected add-ons (including multiple quantities of a single add-on)
@@ -64,21 +60,4 @@ fun checkoutProperties(
         .bonusAmount(bonus)
         .shippingAmount(shippingAmount)
         .build()
-}
-
-fun <T : Relay> encodeRelayId(relay: T): String {
-    val classSimpleName = relay.javaClass.simpleName.replaceFirst("AutoParcel_", "")
-    val id = relay.id()
-    return Base64Utils.encodeUrlSafe(("$classSimpleName-$id").toByteArray(Charset.defaultCharset()))
-}
-
-fun decodeRelayId(encodedRelayId: String?): Long? {
-    return try {
-        String(Base64Utils.decode(encodedRelayId), Charset.defaultCharset())
-            .replaceBeforeLast("-", "", "")
-            .toLong()
-            .absoluteValue
-    } catch (e: Exception) {
-        null
-    }
 }
